Added GUI info to README, reformatted links
Now that there's a decent GUI, we can at least mention that it exists in
the README.  It's good to make people aware of their options, you know?

Also, I got tired of the inline URLs, so I changed the links to be of
the footnote type.  It's much nicer looking e.g. in the list of firmware
flashing programs.