Compile firmware for STM32F072x8 instead of xB
Since the firmware is under 32 KiB, I figured there's not much reason to
use a microcontroller with 128 KiB of flash.  The new boards use
STM32F072x8 chips, which only have 64 KiB of flash.  This allows for the
firmware to grow, while not being quite so ridiculous as before.

This commit sets the Makefile to put the configuration at the end of the
64 KiB flash chip, and to use a new linker script for the same chip.