Better explain how to enter DFU mode
Starting from v0.3, the PD Buddy Sink is sold without SW1 installed.  It
might be frustrating to see the README mention setting the switch only
to find that there isn't actually a switch to set.

Now the README mentions that there might not be a switch, and what to do
if there isn't.  Also, it's more specific about what position for the
switch makes the device enter DFU mode.