Documented some of the firmware changes
The new behavior of `write` has been documented, as well as a high-level
description of the new Setup mode behavior with Power Delivery
communications.

The new `output` command now has documented specifications.  Not a
single line of code has been written for it though, but this is isn't a
release so you shouldn't take everything the documentation file says as
true.