Correctly work with Type-C Current in Setup mode
When we were getting power from Type-C Current advertisements, we used
to not update the location of the configuration object.  This was done
to save time between polling the Type-C Current, and was not a problem
until we started trying to do Power Delivery in Setup mode.  As the
comment correctly said before, when we were trying to get power, the
location of the configuration object wouldn't change, so there was no
need to find its location more than once.

Now that the location of the configuration can change while we're
getting power, we need to check if that's happened.  This can be easily
done by checking the configuration object's status.  If the status is
not VALID, we get the location of the new configuration object.