Remove a bool in pdb_dpm_evaluate_typec_current
The cfg_set variable was not necessary after all, since testing if
cfg == NULL works just as well as testing if !cfg_set.  Actually a bit
better, since it prevents a null dereference we did by mistake before.
I should find out if the null derefernce hurts anything, because if it
does I should make a patch to the 1.1.x series to fix it.