Remember the valid configuration location
Until now, the valid configuration object's location was not stored, so
it had to be looked up every time pdb_config_flash_read was called.  Now
the valid object's location is stored in config_cur, which is updated
every time the configuration's location changes.  This means that
pdb_config_flash_read only has to do the full routine the first time
it's called, or after pdb_config_flash_erase is run.  Now other parts of
the configuration won't have to cache the location themselves anymore.