Only set the output in trans.sink when not min pwr
When we receive a GotoMin message, we immediately go to lower power
before entering the PE_SRC_Transition_Sink state.  Therefore, the call
to pdb_dpm_output_set was redundant in this case.  Now the call isn't
made in this case, removing the redundancy.  This also leads somewhat
into the design of the new library.