Change spec to have one set_vrange command
This commit changes back to the original plan of having one set_vrange
command that sets both the vmin and vmax at once, instead of separate
set_vmin and set_vmax commands.  This will allow checking that the range
is valid every time it's set instead of having to wait until
configuration is written to flash.