Partially implement vrange requests
Requests for fixed voltages are now made from the voltage range.  This
isn't a terrible behavior, but I might want to be able to make requests
from programmable supplies from the range as well.  'Twould be more
complicated though, and provide comparitively little benefit.  I'll
think about it, and maybe put out a request for comments.