Update to ChibiOS 17.6.x
Technically I updated to a development version somewhere between 17.6.3
and 17.6.4, but I don't doubt its stability.

This is a good update because I'd like users of the library to be
confident that it works with the latest version of ChibiOS.
Interestingly, none of the changes I made were in the library code, so
it did before, but now I know for sure.

I'm unclear as to whether I can now get rid of the INT_N thread in favor
of a pin interrupt.  I'll have to see if I can, or if that will have to
wait for still more ChibiOS updates.