Automatically run `load` when setup mode starts
Gavin had trouble with configuration because it isn't loaded on startup.
I realized that I've been caught with the same mistake myself before, so
since at least 2 out of ~65 people have done this, it's clearly a real
usability problem.

Now the configuration is automatically loaded from flash when setup mode
starts, so there's no need to type `load` to edit the configuration the
Sink already has.  If you're changing all the settings, you'll have to
type them all anyway, so I don't think there will be any surprises
because of this change.