Add flash-bmp target to Makefile
I have a Black Magic Probe now, so I'd like to be able to use that to
flash the PD Buddy Sink.  This commit adds a Makefile rule to automate
the very long command to do that.  I didn't bother with a GDB script as
the BMP wiki suggests, since it seems to add little advantage when the
command is already hidden away in a Makefile.