Update to ChibiOS 18.2.1
Another big ChibiOS update, this time to the first stable release in the
18.2.x series.  This time a few pieces of API that the firmware and
library used were renamed, so there were some changes that had to be
made there.  The build system is nicer now, as the included makefiles
add paths to the necessary variables so the main Makefile is much less
verbose in a few places.