Added LED status code for default USB power
It's useful to know when only default USB power is available, as opposed
to when some mechanism for negotiating higher power is available but it
isn't possible to get high enough power.  This can let users know e.g.
using a USB A to C cable can't provide high power, potentially reducing
confusion.

The blink code for this is 1 7/8 s on, 1/8 s off.  I think this is
sufficiently unlike other blink patterns that it won't cause any
confusion.